larger output capacity

+++ src/gd_internal.h
@@ -22,7 +22,7 @@ typedef struct {
     uint16_t entries_length;
     gd_string_table_entry_t *entries;
     size_t strings_capacity;
-    uint16_t strings_length;
+    size_t strings_length;
     gd_index_t *strings;
 } gd_string_table_t;
 
@@ -40,8 +40,8 @@ typedef struct {
     gd_string_table_t string_table;
     gd_string_t prior_string;
     gd_index_t *output;
-    uint16_t outputCapacity;
-    uint16_t outputLength;
+    size_t outputCapacity;
+    size_t outputLength;
 } gd_expand_codes_t;
 
 // unpack state variables, needed for subblocks
